+++ luhn.h
@@ -12,12 +12,12 @@ using namespace std;
 class Luhn  // class declaration
 {
 private: 
-	static long get_checksum(unsigned long long account_number, bool contains_check_digit);
+	static unsigned int get_checksum(unsigned long long account_number, bool contains_check_digit);
 	
 public:
     static bool isValid(unsigned long long account_number);
-  	static long get_checkdigit(unsigned long long account_number, bool contains_check_digit);
-  	static long get_checkdigit(unsigned long long account_number);
+  	static unsigned int get_checkdigit(unsigned long long account_number, bool contains_check_digit);
+  	static unsigned int get_checkdigit(unsigned long long account_number);
 	static vector<int> string_to_int_vector(string value);
 	static int summarize_int_vector(vector<int> numbers);
 };
